+++ src/bsdd_gui/module/iso_export/trigger.py
@@ -24,7 +24,7 @@ def create_widget(data: object, parent: ui.Widget):
 
 
 def widget_created(widget: ui.Widget):
-    core.register_widget(widget, tool.IsoExport,tool.Loin)
+    core.register_widget(widget, tool.IsoExport,tool.PropertyPicker)
     core.register_fields(widget, tool.IsoExport)
     core.register_validators(widget, tool.IsoExport, tool.Util)
     core.connect_widget(widget, tool.IsoExport)
