--- src/bsdd_gui/module/loin/__init__.py
@@ -1,18 +1,0 @@
-import bsdd_gui
-from . import ui, prop, trigger
-
-
-def register():
-    bsdd_gui.LoinProperties = prop.LoinProperties()
-
-
-def retranslate_ui():
-    trigger.retranslate_ui()
-
-
-def load_ui_triggers():
-    trigger.connect()
-
-
-def on_new_project():
-    trigger.on_new_project()
