--- src/bsdd_gui/module/loin/ui.py
@@ -1,2 +1,0 @@
-# The loin module is a data/service module without its own widget; UI surfaces
-# live in property_picker (UC/MS editor) and iso_export (LOIN XML export).
