+++ src/bsdd_gui/module/property_picker/constants.py
@@ -1,0 +1,4 @@
+# Default seed values for a fresh LOIN — keep one UC and one MS so the
+# property picker always has at least one checkable UC × MS column.
+DEFAULT_PURPOSE_NAME = "Base Use Case"
+DEFAULT_MILESTONE_NAME = "Base Milestone"
