+++ src/bsdd_gui/presets/signal_presets.py
@@ -8,7 +8,11 @@ from bsdd_gui.presets.ui_presets import FieldWidget, BaseDialog, BaseWindow
 import logging
 
 
-class PluginSignals(QObject):
+
+class BaseSignals(QObject):
+    pass
+
+class PluginSignals(BaseSignals):
 
     def get_signals(self) -> dict[str, Signal]:
         return {name: obj for name, obj in self.__dict__.items() if isinstance(obj, Signal)}
@@ -22,7 +26,7 @@ class PluginSignals(QObject):
                 logging.debug(str(e))
 
 
-class WidgetSignals(QObject):
+class WidgetSignals(BaseSignals):
     widget_requested = Signal()
     widget_closed = Signal(BaseWindow)
     widget_created = Signal(BaseWindow)
